02 Scripts/requirements.txt
python-dotenv=0.12.0
piexif=1.1.3

